+++ package.json
@@ -45,7 +45,8 @@
     "tmp": "0.2.1",
     "user-settings": "0.2.0",
     "xlsx": "0.16.4",
-    "xmldom": "^0.5.0"
+    "xmldom": "^0.5.0",
+    "@here/olp-sdk-authentication": "1.11.0"
   },
   "devDependencies": {
     "@types/crypto-js": "^3.1.47",
